+++ src/test/java/com/github/alexisjehan/mvncheck/core/ServiceTest.java
@@ -337,7 +337,8 @@ final class ServiceTest {
 	}
 
 	@Test
-	void testFindArtifactUpdateVersions() throws IOException {
+	@Deprecated
+	void testFindArtifactUpdateVersionsDeprecated() throws IOException {
 		final var service = new Service(
 				Set.of(mockedMavenBuildResolver, mockedGradleBuildResolver),
 				mockedArtifactAvailableVersionsResolver
@@ -393,10 +394,95 @@ final class ServiceTest {
 	}
 
 	@Test
-	void testFindArtifactUpdateVersionsInvalid() throws IOException {
+	@Deprecated
+	void testFindArtifactUpdateVersionsDeprecatedInvalid() throws IOException {
 		final var service = new Service(mockedMavenSession);
 		assertThatNullPointerException()
 				.isThrownBy(() -> service.findArtifactUpdateVersions(null, false));
+	}
+
+	@Test
+	void testFindArtifactUpdateVersions() throws IOException {
+		final var service = new Service(
+				Set.of(mockedMavenBuildResolver, mockedGradleBuildResolver),
+				mockedArtifactAvailableVersionsResolver
+		);
+		final var fooIdentifier = new ArtifactIdentifier("foo-group-id", "foo-artifact-id");
+		final var barIdentifier = new ArtifactIdentifier("bar-group-id", "bar-artifact-id");
+		Mockito
+				.when(
+						mockedArtifactAvailableVersionsResolver.resolve(
+								Mockito.argThat(
+										artifact -> null != artifact && fooIdentifier.equals(artifact.getIdentifier())
+								),
+								Mockito.notNull()
+						)
+				)
+				.then(
+						invocation -> new ArtifactAvailableVersions(
+								invocation.getArgument(0),
+								List.of("1.0.0", "2.0.0")
+						)
+				);
+		Mockito
+				.when(
+						mockedArtifactAvailableVersionsResolver.resolve(
+								Mockito.argThat(
+										artifact -> null != artifact && barIdentifier.equals(artifact.getIdentifier())
+								),
+								Mockito.notNull()
+						)
+				)
+				.then(
+						invocation -> new ArtifactAvailableVersions(
+								invocation.getArgument(0),
+								List.of()
+						)
+				);
+		final var fooArtifact1 = new Artifact<>(
+				MavenArtifactType.DEPENDENCY,
+				fooIdentifier,
+				"1.0.0",
+				true
+		);
+		final var fooArtifact2 = new Artifact<>(
+				MavenArtifactType.DEPENDENCY,
+				fooIdentifier,
+				"2.0.0-SNAPSHOT"
+		);
+		final var fooArtifact3 = new Artifact<>(
+				MavenArtifactType.DEPENDENCY,
+				fooIdentifier,
+				"2.0.0"
+		);
+		final var barArtifact = new Artifact<>(
+				MavenArtifactType.DEPENDENCY,
+				barIdentifier,
+				"1.1.0"
+		);
+		final var build = new Build(
+				new BuildFile(BuildFileType.MAVEN, Path.of("src", "test", "resources", "pom.xml")),
+				List.of(),
+				List.of(fooArtifact1, fooArtifact2, fooArtifact3, barArtifact)
+		);
+		assertThat(service.findArtifactUpdateVersions(build, false, false)).containsExactly(
+				new ArtifactUpdateVersion(fooArtifact1, "2.0.0"),
+				new ArtifactUpdateVersion(fooArtifact2, "2.0.0")
+		);
+		assertThat(service.findArtifactUpdateVersions(build, true, false)).containsExactly(
+				new ArtifactUpdateVersion(fooArtifact2, "2.0.0")
+		);
+		assertThat(service.findArtifactUpdateVersions(build, false, true)).containsExactly(
+				new ArtifactUpdateVersion(fooArtifact1, "2.0.0")
+		);
+	}
+
+	@Test
+	void testFindArtifactUpdateVersionsInvalid() throws IOException {
+		final var service = new Service(mockedMavenSession);
+		assertThatNullPointerException().isThrownBy(
+				() -> service.findArtifactUpdateVersions(null, false, false)
+		);
 	}
 
 	@Test